Move d3d_device_interop to ogl device
