+++ src/accelerator/ogl/util/device.cpp
@@ -34,6 +34,11 @@
 
 #include <SFML/Window/Context.hpp>
 
+#ifdef WIN32
+#include "../../d3d/d3d_device.h"
+#include <GL/wglew.h>
+#endif
+
 #include <boost/asio/deadline_timer.hpp>
 #include <boost/asio/dispatch.hpp>
 #include <boost/asio/spawn.hpp>
@@ -67,12 +72,20 @@ struct device::impl : public std::enable_shared_from_this<impl>
 
     std::wstring version_;
 
+#ifdef WIN32
+    std::shared_ptr<d3d::d3d_device> const d3d_device_;
+    std::shared_ptr<void>                  interop_handle_;
+#endif
+
     io_context                          service_;
     decltype(make_work_guard(service_)) work_;
     std::thread                         thread_;
 
     impl()
         : device_(sf::ContextSettings(0, 0, 0, 4, 5, sf::ContextSettings::Attribute::Core), 1, 1)
+#ifdef WIN32
+        , d3d_device_(d3d::d3d_device::get_device())
+#endif
         , work_(make_work_guard(service_))
     {
         CASPAR_LOG(info) << L"Initializing OpenGL Device.";
@@ -99,6 +112,18 @@ struct device::impl : public std::enable_shared_from_this<impl>
         GL(glBindFramebuffer(GL_FRAMEBUFFER, fbo_));
 
         device_.setActive(false);
+
+#ifdef WIN32
+        if (d3d_device_) {
+            interop_handle_ = std::shared_ptr<void>(wglDXOpenDeviceNV(d3d_device_->device()), [](void* p) {
+                if (p)
+                    wglDXCloseDeviceNV(p);
+            });
+
+            if (!interop_handle_)
+                CASPAR_THROW_EXCEPTION(gl::ogl_exception() << msg_info("Failed to initialize d3d interop."));
+        }
+#endif
 
         thread_ = std::thread([&] {
             device_.setActive(true);
@@ -320,6 +345,7 @@ std::future<array<const uint8_t>> device::copy_async(const std::shared_ptr<textu
     return impl_->copy_async(source);
 }
 #ifdef WIN32
+std::shared_ptr<void>                 device::d3d_interop() const { return impl_->interop_handle_; }
 std::future<std::shared_ptr<texture>> device::copy_async(GLuint source, int width, int height, int stride)
 {
     return impl_->copy_async(source, width, height, stride);
