+++ src/accelerator/ogl/util/device.h
@@ -49,6 +49,7 @@ class device final : public std::enable_shared_from_this<device>
                                       copy_async(const array<const uint8_t>& source, int width, int height, int stride);
     std::future<array<const uint8_t>> copy_async(const std::shared_ptr<class texture>& source);
 #ifdef WIN32
+    std::shared_ptr<void>                 d3d_interop() const;
     std::future<std::shared_ptr<texture>> copy_async(GLuint source, int width, int height, int stride);
 #endif
     template <typename Func>
